requirements.txt
streamlit
pandas
fitz
PyMuPDF
python-docx
scikit-learn
xlsxwriter